Updated the content of the main section of the HTML document
Added the details of the weather for the coming days including the day, the temperature and the forecast
I made use of semantic elements like section and main
Changed the emojis to animations using Embed HTML from Lottiefiles. This enabled me to add my personal style to the project and make it look more appealing.